+++ scripts/.vscode/tasks.json
@@ -1,0 +1,17 @@
+{
+    // See https://go.microsoft.com/fwlink/?LinkId=733558
+    // for the documentation about the tasks.json format
+    "version": "2.0.0",
+    "tasks": [
+        {
+            "label": "mpv",
+            "type": "shell",
+            "command": "/usr/bin/mpv /mnt/smb/Anime/Archive/Yu-Gi-Oh\\!\\ -\\ The\\ Dark\\ Side\\ of\\ Dimensions.mkv",
+            "problemMatcher": [],
+            "group": {
+                "kind": "build",
+                "isDefault": true
+            }
+        }
+    ]
+}
